discord_ai_requirements.txt
discord.py
python-dotenv
requests
openai 